add who_alive subset function for baseline. add subsetting to map aggregates.
